+++ src/UnityFx.DependencyInjection.Tests/Helpers/ArrayListDependentClass.cs
@@ -2,12 +2,13 @@
 // Licensed under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System;
+using System.Collections;
 
 namespace UnityFx.DependencyInjection
 {
-	internal class InvalidService
+	internal class ArrayListDependentClass
 	{
-		public InvalidService(int arg)
+		public ArrayListDependentClass(ArrayList arg)
 		{
 		}
 	}