+++ src/UnityFx.DependencyInjection.Tests/Tests/ServiceProviderTests.cs
@@ -6,18 +6,53 @@ using System.Collections;
 using Xunit;
 using NSubstitute;
 
-namespace UnityFx.AppStates.DependencyInjection
+namespace UnityFx.DependencyInjection
 {
 	public class ServiceProviderTests
 	{
-		[Fact]
-		public void GetService_Throws_ServiceNotFoundException()
+		[Theory]
+		[InlineData(typeof(SelfReferenceClass))]
+		[InlineData(typeof(CrossReferenceClass1))]
+		[InlineData(typeof(CrossReferenceClass2))]
+		[InlineData(typeof(InvalidService))]
+		[InlineData(typeof(ArrayListDependentClass))]
+		public void BuildServiceProvider_ThrowsOnError(Type type)
 		{
 			// Arrange
-			var sp = new ServiceProvider();
+			var sc = new ServiceCollection();
+			sc.AddTransient(type);
 
 			// Act/Assert
-			Assert.Throws<ServiceNotFoundException>(() => sp.GetService(typeof(IEnumerable)));
+			Assert.Throws<InvalidOperationException>(() => sc.BuildServiceProvider());
+		}
+
+		[Fact]
+		public void BuildServiceProvider_ThrowsIfScopedServiceResolvedFromSingleton()
+		{
+			// Arrange
+			var sc = new ServiceCollection();
+			sc.AddSingleton<ArrayListDependentClass>();
+			sc.AddScoped<ArrayList>();
+
+			// Act/Assert
+			Assert.Throws<InvalidOperationException>(() => sc.BuildServiceProvider());
+		}
+
+		[Fact]
+		public void Dispose_IsCalledOnScopeDisposal()
+		{
+			// Arrange
+			var sc = new ServiceCollection();
+			var disposable = Substitute.For<IDisposable>();
+			sc.AddSingleton(disposable);
+
+			var sp = sc.BuildServiceProvider();
+
+			// Act
+			sp.Dispose();
+
+			// Assert
+			disposable.Received(1).Dispose();
 		}
 
 		[Theory]
@@ -25,39 +60,105 @@ namespace UnityFx.AppStates.DependencyInjection
 		[InlineData(typeof(CrossReferenceClass1))]
 		[InlineData(typeof(CrossReferenceClass2))]
 		[InlineData(typeof(InvalidService))]
-		public void GetService_Throws_ServiceConstructorResolutionException(Type type)
+		[InlineData(typeof(ArrayListDependentClass))]
+		public void CreateInstance_ThrowsOnError(Type type)
 		{
 			// Arrange
-			var sp = new ServiceProvider();
-			sp.Add(new ServiceDescriptor(type, type, ServiceLifetime.Transient));
+			var sc = new ServiceCollection();
+			var sp = sc.BuildServiceProvider();
 
 			// Act/Assert
-			Assert.Throws<ServiceConstructorResolutionException>(() => sp.GetService(type));
+			Assert.Throws<InvalidOperationException>(() => sp.CreateInstance(type));
+		}
+
+		[Fact]
+		public void CreateInstance_MatchesArguments()
+		{
+			// Arrange
+			var sc = new ServiceCollection();
+			var sp = sc.BuildServiceProvider();
+
+			// Act
+			var result = sp.CreateInstance<ArrayListDependentClass>(new ArrayList());
+
+			// Assert
+			Assert.NotNull(result);
+		}
+
+		[Fact]
+		public void GetService_ReturnsNullIfServiceNotFound()
+		{
+			// Arrange
+			var sc = new ServiceCollection();
+			var sp = sc.BuildServiceProvider();
+
+			// Act
+			var result = sp.GetService(typeof(IEnumerable));
+
+			// Assert
+			Assert.Null(result);
 		}
 
 		[Fact]
 		public void GetService_CreatesOnlyOneSingletonInstance()
 		{
 			// Arrange
-			var sp = new ServiceProvider();
-			sp.Add(new ServiceDescriptor(typeof(IEnumerable), typeof(ArrayList), ServiceLifetime.Singleton));
+			var sc = new ServiceCollection();
+			sc.AddSingleton<IEnumerable, ArrayList>();
+
+			var sp = sc.BuildServiceProvider();
+			var sp2 = sp.CreateScope().ServiceProvider;
 
 			// Act
 			var instance1 = sp.GetService(typeof(IEnumerable));
 			var instance2 = sp.GetService(typeof(IEnumerable));
+			var instance3 = sp2.GetService(typeof(IEnumerable));
+			var instance4 = sp2.GetService(typeof(IEnumerable));
 
 			// Assert
 			Assert.NotNull(instance1);
 			Assert.NotNull(instance2);
+			Assert.NotNull(instance3);
+			Assert.NotNull(instance4);
 			Assert.Same(instance1, instance2);
+			Assert.Same(instance3, instance4);
+			Assert.Same(instance1, instance3);
+		}
+
+		[Fact]
+		public void GetService_CreatesOneScopedInstancePerScope()
+		{
+			// Arrange
+			var sc = new ServiceCollection();
+			sc.AddScoped<IEnumerable, ArrayList>();
+
+			var sp = sc.BuildServiceProvider();
+			var sp2 = sp.CreateScope().ServiceProvider;
+
+			// Act
+			var instance1 = sp.GetService(typeof(IEnumerable));
+			var instance2 = sp.GetService(typeof(IEnumerable));
+			var instance3 = sp2.GetService(typeof(IEnumerable));
+			var instance4 = sp2.GetService(typeof(IEnumerable));
+
+			// Assert
+			Assert.NotNull(instance1);
+			Assert.NotNull(instance2);
+			Assert.NotNull(instance3);
+			Assert.NotNull(instance4);
+			Assert.Same(instance1, instance2);
+			Assert.Same(instance3, instance4);
+			Assert.NotSame(instance1, instance3);
 		}
 
 		[Fact]
 		public void GetService_CreatesManyTransientInstances()
 		{
 			// Arrange
-			var sp = new ServiceProvider();
-			sp.Add(new ServiceDescriptor(typeof(IEnumerable), typeof(ArrayList), ServiceLifetime.Transient));
+			var sc = new ServiceCollection();
+			sc.AddTransient<IEnumerable, ArrayList>();
+
+			var sp = sc.BuildServiceProvider();
 
 			// Act
 			var instance1 = sp.GetService(typeof(IEnumerable));
@@ -73,8 +174,10 @@ namespace UnityFx.AppStates.DependencyInjection
 		public void GetService_ResolvesMultipleCtors()
 		{
 			// Arrange
-			var sp = new ServiceProvider();
-			sp.Add(new ServiceDescriptor(typeof(MultiCtorClass), typeof(MultiCtorClass), ServiceLifetime.Transient));
+			var sc = new ServiceCollection();
+			sc.AddTransient<MultiCtorClass>();
+
+			var sp = sc.BuildServiceProvider();
 
 			// Act
 			var instance = sp.GetService(typeof(MultiCtorClass));